Fixing yet more

+++ 2020-Robot/src/main/java/frc/team670/robot/subsystems/SparkMaxRotatingSubsystem.java
@@ -27,11 +27,11 @@ public abstract class SparkMaxRotatingSubsystem extends MustangSubsystemBase imp
     protected double kP, kI, kD, kFF, kIz, MAX_OUTPUT, MIN_OUTPUT;
     protected double MAX_VEL, MIN_VEL, MAX_ACC, ALLOWED_ERR;
     protected int SMARTMOTION_SLOT;
-    protected int ROTATOR_GEAR_RATIO;
+    protected double ROTATOR_GEAR_RATIO;
 
     /**
-     * Configuration for this RotatingSubsystem's properties. 
-     * Use this to keep track of PID and SmartMotion constants
+     * Configuration for this RotatingSubsystem's properties. Use this to keep track
+     * of PID and SmartMotion constants
      */
     public static abstract class Config {
 
@@ -41,7 +41,7 @@ public abstract class SparkMaxRotatingSubsystem extends MustangSubsystemBase imp
 
         public abstract MotorConfig.Motor_Type getMotorType();
 
-        public abstract int getRotatorGearRatio();
+        public abstract double getRotatorGearRatio();
 
         public abstract double getP();
 
@@ -128,12 +128,12 @@ public abstract class SparkMaxRotatingSubsystem extends MustangSubsystemBase imp
         rotator_controller.setReference(setpoint, ControlType.kSmartMotion);
     }
 
-    public void setTargetAngleInDegrees(double angle){
-       setSmartMotionTarget(getMotorRotationsFromAngle(angle));
+    public void setTargetAngleInDegrees(double angle) {
+        setSmartMotionTarget(getMotorRotationsFromAngle(angle));
     }
 
-    protected double getMotorRotationsFromAngle(double angle){
-        return (angle/360)*this.ROTATOR_GEAR_RATIO;
+    protected double getMotorRotationsFromAngle(double angle) {
+        return (angle / 360) * this.ROTATOR_GEAR_RATIO;
     }
 
     public abstract double getCurrentAngleInDegrees();
